src/requirements.txt
python-telegram-bot==20.7
SQLAlchemy
python-dotenv
requests